requirements.txt
Flask
python-pptx
flask-mail
python-dotenv
pytest